// event-share-app/src/models/newEvent.ts
interface newEvent {
    Event_Name: string,
    Target_Audience: string,
    Date: string,
    Event_Start_Time: string,
    Event_End_Time: string
    Room_Number: string,
    Location: string,
    Description: string
    Category: string
    Event_Link: string
    Organizer: string
    Price: string
    Save: string
}

export default newEvent;